+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - uploads now return an error instead of panicking and baker deals with it [#27](https://github.com/AdRoll/baker/pull/27)
 - general: replace `${KEY}` in the TOML conf with the `$KEY` env var [#24](https://github.com/AdRoll/baker/pull/24)
 - input: add KCL input. [#36](https://github.com/AdRoll/baker/pull/36)
+- filter: add RegexMatch filter. [#37](https://github.com/AdRoll/baker/pull/37)
 
 ### Changed